Gluvia-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
pydantic
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-asyncio
requests
coverage
pytest-cov
python-dotenv
Pillow
pytesseract
email-validator
httpx
psycopg2
psutil
pymupdf
openai